Time: 78 ms (15.65%), Space: 38.6 MB (90.81%) - LeetHub

+++ 1509-minimum-difference-between-largest-and-smallest-value-in-three-moves/1509-minimum-difference-between-largest-and-smallest-value-in-three-moves.cpp
@@ -5,7 +5,7 @@ public:
         
         int N = nums.size();
         sort(nums.begin(), nums.end());
-        long long answer = nums[N - 1] - nums[0];
+        long answer = nums[N - 1] - nums[0];
         
         for(int i = 0; i <= 3; i++) {
             if(answer > abs(nums[0 + i] - nums[N - 1 - (3 - i)])) {
